+++ src/components/burger-ingredients/tabs/tabs.module.css
@@ -1,0 +1,3 @@
+.tabs {
+  display: flex;
+}
